% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Dada2_processing_functions.R
\name{agglomerate}
\alias{agglomerate}
\title{Agglomeration}
\usage{
agglomerate(df, df_tax, agg_class)
}
\arguments{
\item{df}{numeric ASV/OTU table dataframe with ASVs on rows as row names and samples on columns as column names}

\item{df_tax}{taxonomy table with ASVs on rows as rownames (matching \code{df}).}

\item{agg_class}{character vector of column name in \code{df_tax} by which to agglomerate.}
}
\value{
a dataframe with merged ASV counts and agglomerated taxonomy as rownames.
}
\description{
Takes taxonomy and ASV/OTU table, and merges ASVs which share an indicated taxonomic classification.
}
\examples{
Theis_ASV <- Theis_merged_clean[,9:ncol(Theis_merged_clean)]
Theis_TAX <- Theis_merged_clean[,1:8]
Theis_agg <- agglomerate(df = Theis_ASV, df_tax = Theis_TAX, agg_class = "Genus")
}
