% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Dada2_processing_functions.R
\name{barplot_prep}
\alias{barplot_prep}
\title{Preparation for Barplotting}
\usage{
barplot_prep(df, sample_types)
}
\arguments{
\item{df}{dataframe with genera as columns and samples as rows.}

\item{sample_types}{named list where names are sample types in order, and values are number of samples per type.}
}
\value{
Produces a taxa by sample type dataframe.
}
\description{
Given a dataframe with genera as columns and samples as rows, this will produce a taxa by sample type dataframe.
}
\examples{

types_list <- list("Amnion-Chorion" = 28, "Control" = 41, "Villous Tree" = 29)
prebar <- barplot_prep(preheat_ab, sample_types = types_list)
}
