% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Dada2_processing_functions.R
\name{blaster}
\alias{blaster}
\title{Blaster}
\usage{
blaster(seqs, names, alignments = 1)
}
\arguments{
\item{seqs}{a character vector of 16S rRNA gene sequences to be cross checked against a local 16S rRNA gene database}

\item{names}{a character vector equal in length to \code{seqs} which denotes the query name for each sequence.}

\item{alignments}{an integer value designating how many alignments should be returned per sequence query.}
}
\value{
A summary table listing the query names, Expect value, Bitscore, Subject TAX ID, Percent Identical, Query Coverage, and Subject Title.
}
\description{
Sequence queries are crosschecked to a local 16S rRNA gene database. Subject titles returned are the names of the database sequences which match the query sequences.
This function needs to run in the top level project directory i.e. /Analysis_Ready_Files.
}
\examples{
Blast_results <- blaster(Theis_merged_clean$X[1:2], names = c("ASV1", "ASV2"), alignments = 5)
}
