% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Dada2_processing_functions.R
\name{col_reorder}
\alias{col_reorder}
\title{Column Reorder}
\usage{
col_reorder(dt)
}
\arguments{
\item{dt}{input which is a dada2 taxa table fused to the ASV table.}
}
\value{
Reordered data table
}
\description{
Reorders numeric columns alphabetically based on sample names.
}
\examples{
merged_clean_reordered <- col_reorder(merged_clean)
}
