% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Dada2_processing_functions.R
\name{combine_V4}
\alias{combine_V4}
\title{Combine V4}
\usage{
combine_V4(ASV_datasets, TAX_datasets, META_datasets)
}
\arguments{
\item{ASV_datasets}{list of ASV tables to combine.}

\item{TAX_datasets}{list of TAX tables to combine.}

\item{META_datasets}{list of META tables to combine.}
}
\value{
a list object which contains the merged ASV count table, metadata, taxonomy, and a map to trace ASV renaming.
}
\description{
Combines sequence data from multiple studies based on ASV sequences from the same variable region. Taxonomy and metadata are also merged.
ASV count tables should have taxa on rows and samples on columns. Rownames should be numbered ASVs identical to rownames in taxonomy.
Taxonomy should include a column labeled \code{ASV} with ASVs numbered e.g. \code{ASV1}.
Metadata columns for each study should match and should not be of type factor.
}
\examples{
Theis_split1_ASV <- Theis_rare_ASV[,
c(rep(TRUE, length(Theis_rare_META)/2), rep(FALSE, length(Theis_rare_META)/2))]
Theis_split2_ASV <- Theis_rare_ASV[,
c(rep(FALSE, length(Theis_rare_META)/2), rep(TRUE, length(Theis_rare_META)/2))]
Theis_split1_ASV <- Theis_split1_ASV[rowSums(Theis_split1_ASV) > 0,]
Theis_split2_ASV <- Theis_split2_ASV[rowSums(Theis_split2_ASV) > 0,]
Theis_split1_META <- Theis_rare_META[
c(rep(TRUE, length(Theis_rare_META)/2), rep(FALSE, length(Theis_rare_META)/2)),]
Theis_split2_META <- Theis_rare_META[
c(rep(FALSE, length(Theis_rare_META)/2), rep(TRUE, length(Theis_rare_META)/2)),]
Theis_rare_TAX <- cbind(data.frame(
ASV = rownames(Theis_rare_TAX), stringsAsFactors = FALSE),
Theis_rare_TAX)
Theis_split1_TAX <- Theis_rare_TAX[
match(x = rownames(Theis_split1_ASV), table = rownames(Theis_rare_TAX)),]
Theis_split2_TAX <- Theis_rare_TAX[
match(x = rownames(Theis_split2_ASV), table = rownames(Theis_rare_TAX)),]
combined <- combine_V4(ASV_datasets = list(Theis_split1_ASV, Theis_split2_ASV),
TAX_datasets = list(Theis_split1_TAX, Theis_split2_TAX),
META_datasets = list(Theis_split1_META, Theis_split2_META))
}
