% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Dada2_processing_functions.R
\name{convert_merged}
\alias{convert_merged}
\title{Convert Merged Dataset}
\usage{
convert_merged(merged, char_range = 1:8, num_range)
}
\arguments{
\item{merged}{Column bound taxonomy with ASVs/OTUs on rows to ASV/OTU counts with samples on columns.}

\item{char_range}{numeric vector giving column range for taxonomy. Default is \code{1:8}}

\item{num_range}{numeric vector giving range for ASV/OTU count table. Default is all other columns to the right of taxonomy.}
}
\value{
merged dataset with taxonomy columns as character and ASV/OTU counts per sample as numeric.
}
\description{
Converts a merged dataset taxonomy columns to character and ASV/OTU counts to numeric to ensure that all columns are of the proper type.
}
\examples{
Theis_converted <- convert_merged(merged = Theis_merged, char_range = 1:8)
}
