% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Dada2_processing_functions.R
\name{decontam_histo_prev_plots}
\alias{decontam_histo_prev_plots}
\title{Generate DECONTAM Histogram and Prevalence Plots}
\usage{
decontam_histo_prev_plots(physeq, thresh = 0.5, study_name)
}
\arguments{
\item{physeq}{phyloseq object with OTU, TAX, and SAM components}

\item{thresh}{DECONTAM score threshold. Default is 0.5}

\item{study_name}{character vector name of the study/title for plot.}
}
\value{
returns a list object of the histogram and prevalence plots.
}
\description{
To be used on a Phyloseq object to generate a histogram of DECONTAM scores and a prevalence plot.
}
\examples{
Theis_META_order$Deco_type <- dplyr::case_when(Theis_META_order$Type == "Placenta" ~ "sample",
Theis_META_order$Type == "Technical Control" ~ "control")
DECO_prep <- decontam_prep(df = Theis_merged_clean,
meta = Theis_META_order,
type = "Deco_type",
sample_col = "Run")
decontam_plots <- decontam_histo_prev_plots(physeq = DECO_prep,
thresh = 0.1,
study_name = "Theis")
}
