% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Dada2_processing_functions.R
\name{decontam_prep}
\alias{decontam_prep}
\title{Preparation for Decontam}
\usage{
decontam_prep(df, type)
}
\arguments{
\item{df}{merged_clean dataset}

\item{type}{a named list where each sample is either "sample" or "control" (as in negative control).}
}
\value{
returns a phyloseq object which can then be used in decontam_histo_prev_plots()
}
\description{
Prepares a merged_clean dataset that has been run through dadaset_clean() and col_reorder() for running Decontam.
}
\examples{
decontam_types <- list("sample" = 28, "control" = 41, "sample" = 29)
pre_decontam <- decontam_prep(df = merged_clean, type = decontam_types)
}
