% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Dada2_processing_functions.R
\name{decontaminate}
\alias{decontaminate}
\title{Decontaminate Dataset}
\usage{
decontaminate(df, physeq, thresh)
}
\arguments{
\item{df}{dataframe from which the phyloseq object was generated with decontam_prep()}

\item{physeq}{phyloseq object created from running decontam_prep()}

\item{thresh}{threshold chosen from looking at histogram and prevalence plots}
}
\value{
gives a list of dataframes split into true taxa, contaminants, and the chosen threshold.
}
\description{
Final function for using Decontam.
}
\examples{
decontam_types <- list("sample" = 28, "control" = 41, "sample" = 29)
pre_decontam <- decontam_prep(df = merged_clean, type = decontam_types)
decontaminated <- decontaminate(merged_clean, pre_decontam, thresh = 0.1)
true_taxa_0.1 <- decontaminated$TrueTaxa; contaminants_0.1 <- decontaminated$Contaminants
}
