% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Dada2_processing_functions.R
\name{dephy}
\alias{dephy}
\title{DePhyloseqize}
\usage{
dephy(phyloseq_obj)
}
\arguments{
\item{phyloseq_obj}{the phyloseq object to be converted}
}
\value{
a merged taxonomy and ASV/OTU table as a data frame.
}
\description{
Takes a phyloseq object and converts it back into a merged taxonomy and ASV/OTU table.
}
\examples{
phy <- phyloseqize(merged_df = Theis_merged_clean, taxa_as_rows = TRUE, keep_ASV_nums = TRUE)
merged_dephy <- dephy(phyloseq_obj = phy)
}
