% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Dada2_processing_functions.R
\name{get_top_ASVs}
\alias{get_top_ASVs}
\title{Get Top ASVs}
\usage{
get_top_ASVs(df, number, ASVs_on_Rows = TRUE)
}
\arguments{
\item{df}{numeric ASV/OTU table.}

\item{number}{numeric vector indicating how many top ASVs to return.}

\item{ASVs_on_Rows}{boolean vector indicating whether ASVs in \code{df} are on rows (\code{TRUE}) or columns (\code{FALSE})}
}
\value{
character vector of the top ASVs/OTUs in a dataset.
}
\description{
A simple function to return the names of the top ASVs in a dataset.
}
\examples{
Theis_topASVs <- get_top_ASVs(df = Theis_rare_ASV, number = 10, ASVs_on_Rows = TRUE)
}
