% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Dada2_processing_functions.R
\name{heatmap_prep}
\alias{heatmap_prep}
\title{Preparation for Heatmapping}
\usage{
heatmap_prep(df, taxa_cutoff = 0, mean_ab_cutoff = TRUE)
}
\arguments{
\item{df}{dataset in merged format}

\item{taxa_cutoff}{how many taxa to keep. Default is zero}

\item{mean_ab_cutoff}{whether or not to apply a 1\% mean abundance cutoff.}
}
\value{
returns a dataset ready to input into a heatmapping function.
}
\description{
Takes a cleaned merged dataset from dadaset_clean() and prepares the dataset for heatmapping. Only one cutoff metric should be used at a time.
}
\examples{
preheat_top_10 <- heatmap_prep(merged_clean, taxa_cutoff = 10, mean_ab_cutoff = FALSE)
preheat_mean_ab <- heatmap_prep(merged_clean, taxa_cutoff = 0, mean_ab_cutoff = TRUE)
}
