% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Dada2_processing_functions.R
\name{heatmapping}
\alias{heatmapping}
\title{Heatmapping}
\usage{
heatmapping(df, scalecolor, types, title, by_percent = TRUE)
}
\arguments{
\item{df}{dataset after running heatmap_prep()}

\item{scalecolor}{color theme to be used for the heatmap. Options are "red" or "blue"}

\item{types}{sample types as a named list.}

\item{title}{title for heatmap}

\item{by_percent}{logical. If true, taxa in heatmap are shown as percentage of samples. If false, taxa are shown as absolute abundances.}
}
\value{
returns a heatmap generated from library(ComplexHeatmap)
}
\description{
Generates a heatmap of a merged dataset after processing with heatmap_prep()
}
\examples{
types_list <- list("Amnion-Chorion" = 28, "Control" = 41, "Villous Tree" = 29)
preheat_top_10 <- heatmap_prep(merged_clean, taxa_cutoff = 10, mean_ab_cutoff = FALSE)
heatmapping(df = preheat_top_10, scalecolor = "blue", types = types_list, title = "Example Heatmap")
}
