% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Dada2_processing_functions.R
\name{JS_group_div}
\alias{JS_group_div}
\title{Jensen-Shannon Sample Clustering by Group}
\usage{
JS_group_div(df, meta, Grouping_Col)
}
\arguments{
\item{df}{an ASV/OTU table with samples as column names and ASV/OTUs as row names. The entire dataframe should be ASV/OTU counts, not taxonomy.}

\item{meta}{an associated metadata file which is ordered by the order of samples in the ASV/OTU table, and which has a column to group samples by.}

\item{Grouping_Col}{character vector of the column name by which to group by.}
}
\value{
A dataframe with group names, and mean Jensen-Shannon divergences.
}
\description{
Calculates the average Jensen-Shannon Divergence within sample groups. Can be used to determine degree of similarity.
}
\examples{
JS_clusters <- JS_group_div(df = t(Theis_rare_ASV),
meta = Theis_rare_META,
Grouping_Col = "Type")
}
