% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Dada2_processing_functions.R
\name{neg_taxa_remove}
\alias{neg_taxa_remove}
\title{Remove Control Taxa}
\usage{
neg_taxa_remove(df, control_sample_columns)
}
\arguments{
\item{df}{merged and cleaned dataset to be worked on}

\item{control_sample_columns}{numeric indices of negative control columns.}
}
\value{
returns a dataset without taxa found in negative controls.
}
\description{
Removes taxa found in negative control samples.
}
\examples{
merged_no_control_taxa <- neg_taxa_remove(merged_clean, control_sample_columns = 37:77)
}
