% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Dada2_processing_functions.R
\name{phyloseqize}
\alias{phyloseqize}
\title{Phyloseqize}
\usage{
phyloseqize(merged_df, tax_df, taxa_as_rows = TRUE, keep_ASV_nums = TRUE)
}
\arguments{
\item{merged_df}{a numeric dataframe of ASV sample counts. Can include taxonomy as well. \code{merged_df} and \code{tax_df} row names should be identical.}

\item{tax_df}{optional, if taxonomy is not included in \code{merged_df} include taxonomy table. For best results, create a taxonomy data frame with ASV row names and taxonomic levels as column names. Then convert to a matrix to use as \code{tax_df}.}

\item{taxa_as_rows}{boolean vector indicating whether ASVs are on rows or on columns. Default is \code{TRUE}}

\item{keep_ASV_nums}{boolean vector indicating whether to keep row names of ASV/OTU table.}
}
\value{
a phyloseq object with ASV/OTU count data and associated taxonomy.
}
\description{
Convert data in R to a phyloseq object
}
\examples{
phy <- phyloseqize(merged_df = Theis_merged_clean, taxa_as_rows = TRUE, keep_ASV_nums = TRUE)
}
