% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{preheat_ab}
\alias{preheat_ab}
\title{Data Set from Theis 2019 "Does the Human Placenta..." Prepared for Heatmapping}
\format{
A data frame with 98 rows and 15 variables:
\describe{
\item{Achromobacter}{Taxa Genus Classification 1}
\item{Delftia}{Taxa Genus Classification 2}
\item{Phyllobacterium}{Taxa Genus Classification 3}
\item{Stenotrophomonas}{Taxa Genus Classification 4}
\item{Blastomonas}{Taxa Genus Classification 5}
\item{Clostridium_sensu_stricto_5}{Taxa Genus Classification 6}
\item{Methylobacterium-Methylorubrum}{Taxa Genus Classification 7}
\item{Methylobacterium-Methylorubrum.1}{Taxa Genus Classification 8}
\item{0319-6G20 Unclassified}{Taxa Genus Classification 9}
\item{Ralstonia}{Taxa Genus Classification 10}
\item{Leucobacter}{Taxa Genus Classification 11}
\item{Bacillus}{Taxa Genus Classification 12}
\item{Acinetobacter.1}{Taxa Genus Classification 13}
\item{Leucobacter.1}{Taxa Genus Classification 14}
\item{Cutibacterium}{Taxa Genus Classification 15}
}
}
\source{
Data set prepared for heatmapping.
}
\usage{
preheat_ab
}
\description{
16s rRNA sequencing data processed using DADA2tools functions.
Genera are on columns and Samples are on rows.
}
\keyword{datasets}
