% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Dada2_processing_functions.R
\name{prev_write_xlsx}
\alias{prev_write_xlsx}
\title{Create Excel Prevalence Workbook}
\usage{
prev_write_xlsx(objects, sheetnames, wb_name = "Prevalence")
}
\arguments{
\item{objects}{Objects as a list i.e. list(prev_x, prev_y, prev_z)}

\item{sheetnames}{Names for sheets of workbook i.e. c("Prev X", "Prev Y", "Prev Z")}

\item{wb_name}{Workbook Name}
}
\value{
creates an excel workbook with all prevalence objects as their own sheet.
}
\description{
Combines prevalence objects into different sheets of an excel workbook.
}
\examples{
Prev_AC <- prevalence(merged_clean, subset = TRUE, lt_col = 9, rt_col = 36)
Prev_C <- prevalence(merged_clean, subset = TRUE, lt_col = 37, rt_col = 77)
Prev_V <- prevalence(merged_clean, subset = TRUE, lt_col = 78, rt_col = 106)
Prevs <- list(Prev_AC, Prev_C, Prev_V)
prev_write_xlsx(objects = Prevs, sheetnames = c("Prev AC", "Prev Ctrl", "Prev V"), wb_name = "Prev")
}
