% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Dada2_processing_functions.R
\name{prevalence}
\alias{prevalence}
\title{Taxa Prevalence Determination}
\usage{
prevalence(df, subset = FALSE, lt_col, rt_col)
}
\arguments{
\item{df}{merged_clean dataset or other dataset where taxa are on rows and samples are on columns.}

\item{subset}{Logical. If true, must include lt_col and rt_col parameters.}

\item{lt_col}{leftmost column of subset. Can be numeric or name}

\item{rt_col}{rightmost column of subset. Can be numeric or name}
}
\value{
returns prevalence of taxa in descending order.
}
\description{
Determines taxa prevalence for a subset of a merged dataset.
}
\examples{
Taxa_Prev <- prevalence(merged_clean, subset = TRUE, lt_col = 9, rt_col = NCOL(merged_clean))
}
