% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Dada2_processing_functions.R
\name{row_sample_elim}
\alias{row_sample_elim}
\title{Eliminate Samples in Rows}
\usage{
row_sample_elim(df, thresh = 1)
}
\arguments{
\item{df}{dataset of type data frame with samples on rows and taxa in columns.}

\item{thresh}{Read threshold. Default is 1 i.e. samples with no reads are eliminated.}
}
\value{
returns a dataframe in which all samples remaining have reads.
}
\description{
Checks samples to verify that they have at least a certain number of reads left. Those that don't pass are eliminated.
}
\examples{
Theis_preheat_pass <- row_sample_elim(Theis_preheat_ab, thresh = 1)
}
