% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{sra_key}
\alias{sra_key}
\title{Sample Key for Theis 2019 "Does the Human Placenta..."}
\format{
A dataframe with 242 rows and 4 variables:
\describe{
\item{Run}{SRR number given to each sample}
\item{Host}{Sample origin}
\item{Sample.Name}{Entered sample name}
\item{New_Name}{Optimized sample name for processing with dada2tools}
}
}
\source{
NCBI website SRA Run Selector Metadata
\url{https://www.ncbi.nlm.nih.gov/Traces/study/?acc=PRJNA397876&o=acc_s\%3Aa}
}
\usage{
sra_key
}
\description{
Select columns from the sample metadata.
}
\keyword{datasets}
