% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{Theis_merged}
\alias{Theis_merged}
\title{Data set from Theis 2019 "Does the Human Placenta..."}
\format{
A dataframe with 1925 rows and 250 variables:
\describe{
\item{X}{Sequence from 16s rRNA region of Taxa genome}
\item{Kingdom}{Kingdom Classification of ASV}
\item{Phylum}{Phylum Classification of ASV}
\item{Class}{Class Classification of ASV}
\item{Order}{Order Classification of ASV}
\item{Family}{Family Classification of ASV}
\item{Genus}{Genus Classification of ASV}
\item{Species}{Species Classification of ASV}
\item{SRR number}{NCBI Sample number}
}
}
\source{
Files from NCBI website SRA Run Selector
\url{https://www.ncbi.nlm.nih.gov/Traces/study/?acc=PRJNA397876&o=acc_s\%3Aa}
}
\usage{
Theis_merged
}
\description{
16s rRNA sequencing data processed using Dada2.
Placental samples are included with blank and control samples.
}
\keyword{datasets}
