% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{Theis_merged_clean}
\alias{Theis_merged_clean}
\title{Cleaned up Data Set from Theis 2019 "Does the Human Placenta..."}
\format{
A dataframe with 967 rows and 106 variables:
\describe{
\item{X}{Sequence from 16s rRNA region of Taxa genome}
\item{Kingdom}{Kingdom Classification of ASV}
\item{Phylum}{Phylum Classification of ASV}
\item{Class}{Class Classification of ASV}
\item{Order}{Order Classification of ASV}
\item{Family}{Family Classification of ASV}
\item{Genus}{Genus Classification of ASV}
\item{Species}{Species Classification of ASV}
\item{Sample Name}{Descriptive sample name.}
}
}
\source{
Filtered and cleaned merged data set.
}
\usage{
Theis_merged_clean
}
\description{
16s rRNA sequencing data processed using Dada2.
Placental samples (AC and V samples) are included with control samples.
}
\keyword{datasets}
