% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{Theis_META}
\alias{Theis_META}
\title{Sample Metadata for Theis 2019 "Does the Human Placenta..."}
\format{
A dataframe with 242 rows and 4 variables:
\describe{
\item{Run}{SRR number given to each sample}
\item{Host}{Sample origin}
\item{Sample.Name}{Entered sample name}
}
}
\source{
NCBI website SRA Run Selector Metadata
\url{https://www.ncbi.nlm.nih.gov/Traces/study/?acc=PRJNA397876&o=acc_s\%3Aa}
}
\usage{
Theis_META
}
\description{
All columns from the sample metadata.
}
\keyword{datasets}
