% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{Theis_rare_ASV}
\alias{Theis_rare_ASV}
\title{Data set from Theis 2019 "Does the Human Placenta..."}
\format{
A dataframe with 201 rows and 99 variables:
\describe{
\item{SRR number}{NCBI Sample number}
}
}
\source{
Files from NCBI website SRA Run Selector
\url{https://www.ncbi.nlm.nih.gov/Traces/study/?acc=PRJNA397876&o=acc_s\%3Aa}
}
\usage{
Theis_rare_ASV
}
\description{
16s rRNA sequencing data processed using Dada2.
Placental samples are included with blank and control samples.
Samples normalized using \code{phyloseq::rarefy_even_depth()}.
}
\keyword{datasets}
