% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{Theis_rare_META}
\alias{Theis_rare_META}
\title{Sample Metadata for Theis 2019 "Does the Human Placenta..."}
\format{
A dataframe with 99 rows and 7 variables:
\describe{
\item{Run}{SRR number given to each sample}
\item{Sample}{Entered sample name}
\item{Type}{Sample type}
\item{GA}{Gestational age}
\item{Delivery}{Placental sample delivery method}
\item{Del_GA}{Delivery method followed by gestational age}
\item{Level}{Placental region sampled}
}
}
\source{
NCBI website SRA Run Selector Metadata
\url{https://www.ncbi.nlm.nih.gov/Traces/study/?acc=PRJNA397876&o=acc_s\%3Aa}
}
\usage{
Theis_rare_META
}
\description{
Select rows of sample metadata for samples run with 30 cycles of PCR after normalization.
}
\keyword{datasets}
