% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Dada2_processing_functions.R
\name{unclassified_replace}
\alias{unclassified_replace}
\title{Replace Unclassified Genus}
\usage{
unclassified_replace(Family, Genus, ...)
}
\arguments{
\item{Family}{Family column of merged dataset}

\item{Genus}{Genus column of merged dataset}

\item{...}{For additional arugments}
}
\value{
Returns Genus column with replaced NA's
}
\description{
This function is used in dadaset_clean to replace NA's with "{Family} Unclassified"
}
