% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Dada2_processing_functions.R
\name{WA_labels}
\alias{WA_labels}
\title{Weighted Average Genus Labels}
\usage{
WA_labels(
  Plot_object,
  ASV_table,
  TAX_table,
  ASVs_to_plot,
  Tax_col,
  ASV_with_Genus = TRUE,
  ASVs_on_rows = TRUE,
  color,
  Greater_than_1per = FALSE,
  invert_x = FALSE,
  invert_y = FALSE
)
}
\arguments{
\item{Plot_object}{object returned from \code{beta_div()}}

\item{ASV_table}{numeric ASV_table dataframe used to produce beta diversity plot. ASVs can be on rows or columns but should be indicated with \code{ASVs_on_rows}}

\item{TAX_table}{dataframe of taxonomy matching ASVs in \code{ASV_table}. Taxa should be on rows.}

\item{ASVs_to_plot}{a character vector of ASV numbers to plot}

\item{Tax_col}{chracter vector of column name in \code{TAX_table} to use for labels if \code{ASV_with_Genus} is set to \code{FALSE}}

\item{ASV_with_Genus}{a boolean vector to determine if ASV numbers should also be plotted along with genus classifications of ASVs.}

\item{ASVs_on_rows}{a boolean vector indicating if ASVs are on rows in \code{ASV_table}}

\item{color}{a character vector of color to plot points.}

\item{Greater_than_1per}{optional boolean vector. If \code{TRUE}, calculates ASVs to plot based on a cutoff of greater than 1\% mean relative abundance.}

\item{invert_x}{optional boolean vector indicating whether to invert points over y-axis. Default is FALSE.}

\item{invert_y}{optional boolean vector indicating whether to invert points over x-axis. Default is FALSE}
}
\value{
Does not return an object, only adds points and labels to a current beta diversity plot.
}
\description{
Plots weighted average genus labels over a beta diversity plot generated by \code{beta_div()}.
}
\examples{
Theis_topASVs <- get_top_ASVs(Theis_rare_ASV, number = 10)
Theis_Beta_plot_PTCWA <- beta_div(df = Theis_rare_ASV,
df_meta = Theis_rare_META,
type = "Type",
study = "Example Plot",
all_O = TRUE,
taxa_on_rows = TRUE,
size = 2,
legendyn = FALSE)
WA_labels(Plot_object = Theis_Beta_plot_PTCWA,
ASV_table = Theis_rare_ASV,
TAX_table = Theis_rare_TAX,
ASVs_on_rows = TRUE,
ASV_with_Genus = FALSE,
Tax_col = "Genus",
color = "grey",
Greater_than_1per = FALSE,
ASVs_to_plot = Theis_topASVs)
}
