% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Dada2_processing_functions.R
\name{barplotting}
\alias{barplotting}
\title{Barplotting}
\usage{
barplotting(df, study, rt_mar = 10)
}
\arguments{
\item{df}{dataframe prepared with barplot_prep()}

\item{study}{name of the study or origin of the data}

\item{rt_mar}{adjustable right margin parameter. For when taxa names exceed space. Default is 10.}
}
\value{
returns a list of barplots.
}
\description{
Function used to create stacked barplots of absolute and relative reads of sample types.
}
\examples{

types_list <- list("Amnion-Chorion" = 28, "Control" = 41, "Villous Tree" = 29)
prebar <- barplot_prep(preheat_ab, sample_types = types_list)
barplotting(df = prebar, study = "Example", rt_mar = 18)
}
