% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Dada2_processing_functions.R
\name{beta_div}
\alias{beta_div}
\title{PCoA plotting}
\usage{
beta_div(
  df,
  sample_types,
  study,
  inset = -0.4,
  rt_margin = 9.1,
  invert_x = FALSE,
  invert_y = FALSE
)
}
\arguments{
\item{df}{a dataset that has been processed by dadaset_clean() and col_reorder()}

\item{sample_types}{a named list of sample types with counts per sample types as values}

\item{study}{name of study from which dataset originated}

\item{inset}{alters the placement of the legend. -0.4 is the default.}

\item{rt_margin}{adjusts the size of the right margin if needed to fit the legend.}

\item{invert_x}{flips plot over the y-axis.}

\item{invert_y}{flips plot over the x-axis.}
}
\value{
returns a Bray-Curtis PCoA plot of sample types with a legend on the right.
}
\description{
Create a PCoA plot based on Bray-Curtis dissimilarity to map samples and visualize similarity.
}
\examples{
types_list <- list("Amnion-Chorion" = 28, "Control" = 41, "Villous Tree" = 29)
beta_div(merged_clean, sample_types = types_list, study = "Example Study")
}
