% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Dada2_processing_functions.R
\name{dadaset_clean}
\alias{dadaset_clean}
\title{Dadaset Clean}
\usage{
dadaset_clean(df, read_thresh = 200)
}
\arguments{
\item{df}{merged dataset}

\item{read_thresh}{Total ASV reads for samples to pass (>= read_thresh)}
}
\value{
returns dataset with only samples that passed.
}
\description{
Clean a merged dataset output from dada2 (include species) (taxa:ASV table)
}
\examples{
merged_clean <- dadaset_clean(merged, read_thresh = 200)
}
