% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Dada2_processing_functions.R
\name{decontam_histo_prev_plots}
\alias{decontam_histo_prev_plots}
\title{Generate Decontam Histogram and Prevalence Plots}
\usage{
decontam_histo_prev_plots(physeq, thresh = 0.5, study_name)
}
\arguments{
\item{physeq}{phyloseq object with OTU, TAX, and SAM components}

\item{thresh}{Decontam score threshold. Default is 0.5}

\item{study_name}{name of the study where the dataset is from.}
}
\value{
returns histogram and prevalence plots.
}
\description{
To be used on a Phyloseq object to generate a histogram of Decontam scores and a prevalence plot.
}
\examples{
decontam_types <- list("sample" = 28, "control" = 41, "sample" = 29)
pre_decontam <- decontam_prep(df = merged_clean, type = decontam_types)
decontam_histo_prev_plots(pre_decontam, thresh = 0.5, study_name = "Study Name")
}
