% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Dada2_processing_functions.R
\name{decontaminate}
\alias{decontaminate}
\title{Decontaminate Dataset}
\usage{
decontaminate(df, physeq, thresh)
}
\arguments{
\item{df}{dataframe from which the phyloseq object was generated with decontam_prep()}

\item{physeq}{phyloseq object created from running decontam_prep()}

\item{thresh}{numeric vector of threshold chosen based on histogram and prevalence plots.}
}
\value{
gives a list of dataframes split into true taxa, contaminants, and the chosen threshold.
}
\description{
Final function for using DECONTAM.
}
\examples{
Theis_META_order$Deco_type <- dplyr::case_when(Theis_META_order$Type == "Placenta" ~ "sample",
Theis_META_order$Type == "Technical Control" ~ "control")
DECO_prep <- decontam_prep(df = Theis_merged_clean,
meta = Theis_META_order, type = "Deco_type", sample_col = "Run")
Theis_decontaminated <- decontaminate(df = Theis_merged_clean, physeq = DECO_prep, thresh = 0.1)
}
