% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Dada2_processing_functions.R
\name{unclassified_replace}
\alias{unclassified_replace}
\title{Unclassified Replace}
\usage{
unclassified_replace(Col_upper, Col_lower, ...)
}
\arguments{
\item{Col_upper}{dataframe column corresponding to higher taxonomic rank.}

\item{Col_lower}{dataframe column corresponding to lower taxonomic rank.}

\item{...}{left blank.}
}
\value{
modified lower taxonomic rank
}
\description{
Auxillary function to \code{dadaset_clean()}. Used in \code{dadaset_clean()} to replace NA's with next highest taxonomy.
}
\examples{
Theis_merged$Genus <- dada2tools:::unclassified_replace(Col_upper = Theis_merged$Family,
Col_lower = Theis_merged$Genus)
}
